// src/stats/types.ts
export interface CollectionHistoricalSale {
  id: string;
  salePriceEth: number;
  timestamp: number;
  tokenImage: string;
  tokenId: string;
  collectionAddress: string;
  collectionName: string;
  chainId: string;
  txHash: string;
  marketplace: string;
  marketplaceAddress: string;
  seller: string;
  buyer: string;
  quantity: string;
}